requirements.txt
google-genai>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiosqlite
websockets
